Use unblob_check all over and remove almost all AssertionErrors.
