+++ decred/decred/dcr/account.py
@@ -4,7 +4,7 @@ Copyright (c) 2020, the Decred developers
 See LICENSE for details
 """
 
-from decred import DecredError
+from decred import DecredError, unblob_check
 from decred.crypto import crypto, opcode
 from decred.util import encode, helpers
 from decred.util.encode import BuildyBytes, ByteArray
@@ -52,30 +52,6 @@ def filterCrazyAddress(addrs):
         list(str): The filtered addresses.
     """
     return [a for a in addrs if a != CrazyAddress]
-
-
-def unblob_check(class_name, version, pushes, check_data):
-    """
-    Check version and pushes to unblob.
-
-    Args:
-        class_name str: the class name that will appear in error messages.
-        version int: the version number that will be checked.
-        pushes int: the number of pushes that will be checked.
-        check_data dict: keys are version numbers, values are number of
-            expected pushes.
-
-    Raises:
-        NotImplementedError if version is not in check_data keys.
-        DecredError if pushes is not the value in check_data keyed by version.
-    """
-    if version not in check_data.keys():
-        raise NotImplementedError(f"{class_name}: unsupported version {version}")
-    expected_pushes = check_data[version]
-    if pushes != expected_pushes:
-        raise DecredError(
-            f"{class_name}: expected {expected_pushes} pushes, got {pushes}"
-        )
 
 
 class KeySource:
